requirements.txt
#image processing:
opencv-python
Pillow
numpy
rembg -> rembg[gpu]
pillow-heif

#api:
fastapi
uvicorn
python-multipart


